Moved K2 connection integration test from Portal to WorkflowCoordinator
