+++ src/Portal/Portal/HttpClients/IWorkflowServiceApiClient.cs
@@ -4,7 +4,6 @@ namespace Portal.HttpClients
 {
     public interface IWorkflowServiceApiClient
     {
-        Task<bool> CheckK2Connection();
         Task<string> GetTaskCurrentStage(string serialNumber);
         Task<bool> ProgressWorkflowInstance(int processId,string serialNumber, string currentTaskStage, string progressToTaskStage);
     }
