+++ src/WorkflowCoordinator/WorkflowCoordinator/HttpClients/WorkflowServiceApiClient.cs
@@ -22,6 +22,26 @@ namespace WorkflowCoordinator.HttpClients
             _httpClient = httpClient;
         }
 
+        public async Task<bool> CheckK2Connection()
+        {
+            var fullUri = new Uri(_uriConfig.Value.K2WebServiceBaseUri, _uriConfig.Value.K2WebServiceGetWorkflowsUri);
+            string data;
+
+            using (var response = await _httpClient.GetAsync(fullUri))
+            {
+                data = await response.Content.ReadAsStringAsync();
+
+                if (!response.IsSuccessStatusCode)
+                    throw new ApplicationException($"StatusCode='{response.StatusCode}'," +
+                                                   $"\n Message= '{data}'," +
+                                                   $"\n Url='{fullUri}'");
+
+            }
+
+            return true;
+
+        }
+
         public async Task<int> CreateWorkflowInstance(int dbAssessmentWorkflowId)
         {
             var fullUri = new Uri(_uriConfig.Value.K2WebServiceBaseUri, _uriConfig.Value.K2WebServiceStartWorkflowInstanceUri + $"/{dbAssessmentWorkflowId}");
